Move MockWebSocket related stuff into separate test helper file
